requirements.txt
# Web scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
webdriver-manager==4.0.1

# Google Drive API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.2
pymupdf==1.23.8

# Database
sqlalchemy==2.0.23

# AI and NLP
openai==1.30.5
langchain==0.0.350
langchain-openai==0.0.2
nltk==3.8.1
spacy==3.7.2

# Data processing
pandas==2.1.3
numpy==1.25.2

# Web framework
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
tqdm==4.66.1 
# PostgreSQL driver
psycopg2-binary==2.9.9
